refactor: move ClassTransformOptions to interfaces folder
