+++ src/interfaces/class-transformer-options.interface.ts
@@ -1,18 +1,4 @@
-/**
- * Allows to specify a map of Types in the object without using @Type decorator.
- * This is useful when you have external classes.
- */
-export interface TargetMap {
-  /**
-   * Target which Types are being specified.
-   */
-  target: Function;
-
-  /**
-   * List of properties and their Types.
-   */
-  properties: { [key: string]: Function };
-}
+import { TargetMap } from './target-map.interface';
 
 /**
  * Options to be passed during transformation.
