+++ src/interfaces/index.ts
@@ -1,1 +1,3 @@
 export * from './class-constructor.type';
+export * from './class-transformer-options.interface';
+export * from './target-map.interface';
